+++ src/algoritmos/ordenacao/GnomeSort.h
@@ -2,6 +2,6 @@
 #define GNOMESORT_H
 
 // Função de ordenação gnome sort
-void gnomeSort(unsigned arr[], unsigned n, unsigned *numComparacoes, unsigned *numTrocas) ;
+void gnomeSort(unsigned arr[], unsigned n, unsigned long long *numComparacoes, unsigned long long *numTrocas) ;
 
 #endif // GNOMESORT_H
