+++ src/algoritmos/ordenacao/InsertionSort.h
@@ -1,6 +1,6 @@
 #ifndef INSERTIONSORT_H
 #define INSERTIONSORT_H
 
-void insertionSort(unsigned *vet, unsigned size, unsigned *numComp, unsigned *numReg);
+void insertionSort(unsigned *vet, unsigned size, unsigned long long *numComp, unsigned long long *numReg);
 
 #endif // INSERTIONSORT_H
