+++ src/algoritmos/ordenacao/quickSortInsertion.h
@@ -1,3 +1,3 @@
 #include <fstream>
 
-void insertionQuickSort(unsigned *a, unsigned inicio, unsigned fim, unsigned m, unsigned *numComparacoes, unsigned *numTrocas);
+void insertionQuickSort(unsigned *a, unsigned inicio, unsigned fim, unsigned m, unsigned long long *numComparacoes, unsigned long long *numTrocas);
